Manage a single interface
